requirements.txt
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.8.5
pornhub-api==0.2.0
hqporner-api==1.0.2
python-telegram-bot==20.4
